+++ packages/config/src/chains.ts
@@ -120,7 +120,11 @@ export const acala = new Parachain({
   nativeAsset: aca,
   parachainId: 2000,
   ss58Format: 10,
-  ws: 'wss://acala-rpc.aca-api.network',
+  ws: [
+    'wss://acala-rpc.aca-api.network',
+    'wss://acala-rpc.dwellir.com',
+    'wss://acala-polkadot.api.onfinality.io/public-ws',
+  ],
 });
 
 export const alphanetAssetHub = new Parachain({
@@ -145,7 +149,7 @@ export const alphanetAssetHub = new Parachain({
   nativeAsset: unit,
   parachainId: 1001,
   ss58Format: 42,
-  ws: 'wss://frag-moonbase-sm-rpc-ws.g.moonbase.moonbeam.network/',
+  ws: ['wss://frag-moonbase-sm-rpc-ws.g.moonbase.moonbeam.network/'],
 });
 
 export const alphanetRelay = new Parachain({
@@ -163,7 +167,7 @@ export const alphanetRelay = new Parachain({
   nativeAsset: unit,
   parachainId: 0,
   ss58Format: 42,
-  ws: 'wss://fro-moon-rpc-1-moonbase-relay-rpc-1.moonbase.ol-infra.network',
+  ws: ['wss://fro-moon-rpc-1-moonbase-relay-rpc-1.moonbase.ol-infra.network'],
 });
 
 export const astar = new Parachain({
@@ -186,7 +190,11 @@ export const astar = new Parachain({
   nativeAsset: astr,
   parachainId: 2006,
   ss58Format: 5,
-  ws: 'wss://rpc.astar.network',
+  ws: [
+    'wss://rpc.astar.network',
+    'wss://astar-rpc.dwellir.com',
+    'wss://astar.api.onfinality.io/public-ws',
+  ],
 });
 
 export const bifrostKusama = new Parachain({
@@ -230,7 +238,12 @@ export const bifrostKusama = new Parachain({
   nativeAsset: bnc,
   parachainId: 2001,
   ss58Format: 6,
-  ws: 'wss://bifrost-rpc.dwellir.com',
+  ws: [
+    'wss://bifrost-rpc.dwellir.com',
+    'wss://us.bifrost-rpc.liebi.com/ws',
+    'wss://bifrost-rpc.liebi.com/ws',
+    'wss://bifrost-parachain.api.onfinality.io/public-ws',
+  ],
 });
 
 export const bifrostPolkadot = new Parachain({
@@ -298,7 +311,12 @@ export const bifrostPolkadot = new Parachain({
   nativeAsset: bnc,
   parachainId: 2030,
   ss58Format: 6,
-  ws: 'wss://bifrost-polkadot-rpc.dwellir.com',
+  ws: [
+    'wss://bifrost-polkadot-rpc.dwellir.com',
+    'wss://eu.bifrost-polkadot-rpc.liebi.com/ws',
+    'wss://hk.p.bifrost-rpc.liebi.com/ws',
+    'wss://bifrost-polkadot.api.onfinality.io/public-ws',
+  ],
 });
 
 export const calamari = new Parachain({
@@ -325,7 +343,10 @@ export const calamari = new Parachain({
   nativeAsset: kma,
   parachainId: 2084,
   ss58Format: 78,
-  ws: 'wss://calamari.systems',
+  ws: [
+    'wss://calamari.systems',
+    // "wss://calamari-rpc.dwellir.com"
+  ],
 });
 
 export const centrifuge = new Parachain({
@@ -345,7 +366,11 @@ export const centrifuge = new Parachain({
   nativeAsset: cfg,
   parachainId: 2031,
   ss58Format: 36,
-  ws: 'wss://fullnode.centrifuge.io',
+  ws: [
+    'wss://fullnode.centrifuge.io',
+    'wss://centrifuge-rpc.dwellir.com',
+    'wss://centrifuge-parachain.api.onfinality.io/public-ws',
+  ],
 });
 
 export const crustShadow = new Parachain({
@@ -369,7 +394,11 @@ export const crustShadow = new Parachain({
   nativeAsset: csm,
   parachainId: 2012,
   ss58Format: 66,
-  ws: 'wss://rpc2-shadow.crust.network',
+  ws: [
+    'wss://rpc2-shadow.crust.network',
+    'wss://rpc-shadow.crust.network/',
+    'wss://rpc-shadow.crustnetwork.app',
+  ],
 });
 
 export const darwinia = new EvmParachain({
@@ -394,7 +423,7 @@ export const darwinia = new EvmParachain({
   parachainId: 2046,
   rpc: 'https://rpc.darwinia.network',
   ss58Format: 18,
-  ws: 'wss://rpc.darwinia.network',
+  ws: ['wss://rpc.darwinia.network', 'wss://darwinia-rpc.dwellir.com'],
 });
 
 export const darwiniaCrab = new EvmParachain({
@@ -419,39 +448,7 @@ export const darwiniaCrab = new EvmParachain({
   parachainId: 2105,
   rpc: 'https://crab-rpc.darwinia.network',
   ss58Format: 18,
-  ws: 'wss://crab-rpc.darwinia.network',
-});
-
-export const equilibrium = new Parachain({
-  assets: [
-    ChainAsset.fromAsset(glmr, {
-      decimals: 9,
-      ids: {
-        id: 1_735_159_154,
-      },
-    }),
-    ChainAsset.fromAsset(eq, {
-      decimals: 9,
-      ids: {
-        id: 25_969,
-      },
-    }),
-    ChainAsset.fromAsset(eqd, {
-      decimals: 9,
-      ids: {
-        id: 6_648_164,
-      },
-    }),
-  ],
-  ecosystem: Ecosystem.Polkadot,
-  genesisHash:
-    '0x89d3ec46d2fb43ef5a9713833373d5ea666b092fa8fd68fbc34596036571b907',
-  key: 'equilibrium',
-  name: 'Equilibrium',
-  nativeAsset: eq,
-  parachainId: 2011,
-  ss58Format: 68,
-  ws: 'wss://node.pol.equilibrium.io',
+  ws: ['wss://crab-rpc.darwinia.network', 'wss://darwiniacrab-rpc.dwellir.com'],
 });
 
 export const hydration = new Parachain({
@@ -561,7 +558,7 @@ export const hydrationAlphanet = new Parachain({
   nativeAsset: hdx,
   parachainId: 2034,
   ss58Format: 63,
-  ws: 'wss://hydradx-moonbase-rpc.play.hydration.cloud',
+  ws: ['wss://hydradx-moonbase-rpc.play.hydration.cloud'],
 });
 
 export const interlay = new Parachain({
@@ -593,7 +590,7 @@ export const interlay = new Parachain({
   nativeAsset: intr,
   parachainId: 2032,
   ss58Format: 2032,
-  ws: 'wss://api.interlay.io/parachain',
+  ws: ['wss://api.interlay.io/parachain', 'wss://interlay-rpc.dwellir.com'],
 });
 
 export const integritee = new Parachain({
@@ -606,7 +603,10 @@ export const integritee = new Parachain({
   nativeAsset: teer,
   parachainId: 2015,
   ss58Format: 13,
-  ws: 'wss://kusama.api.integritee.network',
+  ws: [
+    'wss://kusama.api.integritee.network',
+    'wss://integritee-kusama.api.onfinality.io/public-ws',
+  ],
 });
 
 export const karura = new Parachain({
@@ -640,7 +640,11 @@ export const karura = new Parachain({
   nativeAsset: kar,
   parachainId: 2000,
   ss58Format: 8,
-  ws: 'wss://karura-rpc-0.aca-api.network',
+  ws: [
+    'wss://karura-rpc-0.aca-api.network',
+    'wss://karura-rpc.dwellir.com',
+    'wss://karura.api.onfinality.io/public-ws',
+  ],
 });
 
 export const khala = new Parachain({
@@ -664,7 +668,11 @@ export const khala = new Parachain({
   nativeAsset: pha,
   parachainId: 2004,
   ss58Format: 30,
-  ws: 'wss://khala-rpc.dwellir.com',
+  ws: [
+    'wss://khala-rpc.dwellir.com',
+    'wss://khala-api.phala.network/ws',
+    'wss://khala.api.onfinality.io/public-ws',
+  ],
 });
 
 export const kintsugi = new Parachain({
@@ -690,7 +698,11 @@ export const kintsugi = new Parachain({
   nativeAsset: kint,
   parachainId: 2092,
   ss58Format: 2092,
-  ws: 'wss://api-kusama.interlay.io/parachain',
+  ws: [
+    'wss://api-kusama.interlay.io/parachain',
+    'wss://kintsugi-rpc.dwellir.com',
+    'wss://kintsugi.api.onfinality.io/public-ws',
+  ],
 });
 
 export const kusama = new Parachain({
@@ -703,7 +715,11 @@ export const kusama = new Parachain({
   nativeAsset: ksm,
   parachainId: 0,
   ss58Format: 2,
-  ws: 'wss://kusama-rpc.dwellir.com',
+  ws: [
+    'wss://kusama-rpc.dwellir.com',
+    'wss://kusama.api.onfinality.io/public-ws',
+    'wss://kusama-rpc.polkadot.io',
+  ],
 });
 
 export const kusamaAssetHub = new Parachain({
@@ -734,7 +750,10 @@ export const kusamaAssetHub = new Parachain({
   nativeAsset: ksm,
   parachainId: 1000,
   ss58Format: 2,
-  ws: 'wss://kusama-asset-hub-rpc.polkadot.io',
+  ws: [
+    'wss://asset-hub-kusama-rpc.dwellir.com',
+    'wss://kusama-asset-hub-rpc.polkadot.io',
+  ],
 });
 
 export const litmus = new Parachain({
@@ -754,7 +773,7 @@ export const litmus = new Parachain({
   nativeAsset: lit,
   parachainId: 2106,
   ss58Format: 131,
-  ws: 'wss://rpc.litmus-parachain.litentry.io',
+  ws: ['wss://rpc.litmus-parachain.litentry.io'],
 });
 
 export const mangataKusama = new Parachain({
@@ -780,7 +799,10 @@ export const mangataKusama = new Parachain({
   nativeAsset: mgx,
   parachainId: 2110,
   ss58Format: 42,
-  ws: 'wss://kusama-archive.mangata.online',
+  ws: [
+    'wss://kusama-archive.mangata.online',
+    'wss://kusama-rpc.mangata.online',
+  ],
 });
 
 export const mantaParachain = new Parachain({
@@ -806,7 +828,7 @@ export const mantaParachain = new Parachain({
   nativeAsset: manta,
   parachainId: 2104,
   ss58Format: 77,
-  ws: 'wss://ws.manta.systems',
+  ws: ['wss://ws.manta.systems'],
 });
 
 export const moonbaseAlpha = new EvmParachain({
@@ -918,7 +940,7 @@ export const moonbaseAlpha = new EvmParachain({
   parachainId: 1000,
   rpc: 'https://rpc.api.moonbase.moonbeam.network',
   ss58Format: 1287,
-  ws: 'wss://wss.api.moonbase.moonbeam.network',
+  ws: ['wss://wss.api.moonbase.moonbeam.network'],
 });
 
 export const moonbaseBeta = new EvmParachain({
@@ -966,7 +988,7 @@ export const moonbaseBeta = new EvmParachain({
   parachainId: 888,
   rpc: 'https://frag-moonbase-beta-rpc.g.moonbase.moonbeam.network',
   ss58Format: 1287,
-  ws: 'wss://deo-moon-rpc-1-moonbase-beta-rpc-1.moonbase.ol-infra.network',
+  ws: ['wss://deo-moon-rpc-1-moonbase-beta-rpc-1.moonbase.ol-infra.network'],
 });
 
 export const moonbeam = new EvmParachain({
@@ -1250,7 +1272,7 @@ export const moonbeam = new EvmParachain({
   parachainId: 2004,
   rpc: 'https://rpc.api.moonbeam.network',
   ss58Format: 1284,
-  ws: 'wss://wss.api.moonbeam.network',
+  ws: ['wss://wss.api.moonbeam.network'],
 });
 
 export const moonriver = new EvmParachain({
@@ -1433,27 +1455,7 @@ export const moonriver = new EvmParachain({
   parachainId: 2023,
   rpc: 'https://rpc.api.moonriver.moonbeam.network',
   ss58Format: 1285,
-  ws: 'wss://wss.api.moonriver.moonbeam.network',
-});
-
-export const nodle = new Parachain({
-  assets: [
-    ChainAsset.fromAsset(nodl, {
-      decimals: 11,
-      ids: {
-        id: 'NodleNative',
-      },
-    }),
-  ],
-  ecosystem: Ecosystem.Polkadot,
-  genesisHash:
-    '0x97da7ede98d7bad4e36b4d734b6055425a3be036da2a332ea5a7037656427a21',
-  key: 'nodle',
-  name: 'Nodle',
-  nativeAsset: nodl,
-  parachainId: 2026,
-  ss58Format: 37,
-  ws: 'wss://nodle-rpc.dwellir.com',
+  ws: ['wss://wss.api.moonriver.moonbeam.network'],
 });
 
 export const neuroweb = new Parachain({
@@ -1473,7 +1475,34 @@ export const neuroweb = new Parachain({
   nativeAsset: neuro,
   parachainId: 2043,
   ss58Format: 101,
-  ws: 'wss://neuroweb-rpc.dwellir.com',
+  ws: [
+    'wss://neuroweb-rpc.dwellir.com',
+    'wss://parachain-rpc.origin-trail.network',
+  ],
+});
+
+export const nodle = new Parachain({
+  assets: [
+    ChainAsset.fromAsset(nodl, {
+      decimals: 11,
+      ids: {
+        id: 'NodleNative',
+      },
+    }),
+  ],
+  ecosystem: Ecosystem.Polkadot,
+  genesisHash:
+    '0x97da7ede98d7bad4e36b4d734b6055425a3be036da2a332ea5a7037656427a21',
+  key: 'nodle',
+  name: 'Nodle',
+  nativeAsset: nodl,
+  parachainId: 2026,
+  ss58Format: 37,
+  ws: [
+    'wss://nodle-rpc.dwellir.com',
+    'wss://nodle-rpc.dwellir.com',
+    'wss://nodle-parachain.api.onfinality.io/public-ws',
+  ],
 });
 
 export const originTrailAlphanet = new Parachain({
@@ -1494,7 +1523,7 @@ export const originTrailAlphanet = new Parachain({
   nativeAsset: otp,
   parachainId: 2043,
   ss58Format: 101,
-  ws: 'wss://otp-lunaris-alpha-node-02.origin-trail.network',
+  ws: ['wss://otp-lunaris-alpha-node-02.origin-trail.network'],
 });
 
 export const parallel = new Parachain({
@@ -1517,7 +1546,7 @@ export const parallel = new Parachain({
   nativeAsset: para,
   parachainId: 2012,
   ss58Format: 172,
-  ws: 'wss://parallel-rpc.dwellir.com',
+  ws: ['wss://parallel-rpc.dwellir.com'],
 });
 
 export const peaqAlphanet = new Parachain({
@@ -1551,7 +1580,7 @@ export const peaqAlphanet = new Parachain({
   nativeAsset: agng,
   parachainId: 3013,
   ss58Format: 42,
-  ws: 'wss://moonbeam.peaq.network',
+  ws: ['wss://moonbeam.peaq.network'],
 });
 
 export const peaqChain = new Parachain({
@@ -1615,7 +1644,7 @@ export const peaqChain = new Parachain({
   nativeAsset: peaq,
   parachainId: 3338,
   ss58Format: 42,
-  ws: 'wss://peaq.api.onfinality.io/public-ws',
+  ws: ['wss://peaq.api.onfinality.io/public-ws'],
 });
 
 export const peaqEvmAlphanet = new EvmParachain({
@@ -1656,7 +1685,7 @@ export const peaqEvmAlphanet = new EvmParachain({
   parachainId: 3013,
   rpc: 'https://moonbeam.PEAQ.network',
   ss58Format: 42,
-  ws: 'wss://moonbeam.peaq.network',
+  ws: ['wss://moonbeam.peaq.network'],
 });
 
 export const peaqEvm = new EvmParachain({
@@ -1729,7 +1758,7 @@ export const peaqEvm = new EvmParachain({
   parachainId: 3338,
   rpc: 'https://peaq.api.onfinality.io/public',
   ss58Format: 42,
-  ws: 'wss://peaq.api.onfinality.io/public-ws',
+  ws: ['wss://peaq.api.onfinality.io/public-ws'],
 });
 
 export const pendulum = new Parachain({
@@ -1755,7 +1784,7 @@ export const pendulum = new Parachain({
   nativeAsset: pen,
   parachainId: 2094,
   ss58Format: 56,
-  ws: 'wss://rpc-pendulum.prd.pendulumchain.tech',
+  ws: ['wss://rpc-pendulum.prd.pendulumchain.tech'],
 });
 
 export const pendulumAlphanet = new Parachain({
@@ -1782,7 +1811,7 @@ export const pendulumAlphanet = new Parachain({
   nativeAsset: ampe,
   parachainId: 2124,
   ss58Format: 57,
-  ws: 'wss://moonbeam-00.pendulumchain.tech:443',
+  ws: ['wss://moonbeam-00.pendulumchain.tech:443'],
 });
 
 export const phala = new Parachain({
@@ -1806,7 +1835,12 @@ export const phala = new Parachain({
   nativeAsset: pha,
   parachainId: 2035,
   ss58Format: 30,
-  ws: 'wss://phala-rpc.dwellir.com',
+  ws: [
+    'wss://phala-rpc.dwellir.com',
+    'wss://api.phala.network/ws',
+    'wss://phala.api.onfinality.io/public-ws',
+    'wss://rpc.helikon.io/phala',
+  ],
 });
 
 export const picasso = new Parachain({
@@ -1833,7 +1867,11 @@ export const picasso = new Parachain({
   nativeAsset: pica,
   parachainId: 2087,
   ss58Format: 49,
-  ws: 'wss://picasso-rpc.dwellir.com',
+  ws: [
+    'wss://picasso-rpc.dwellir.com',
+    'wss://rpc.composablenodes.tech',
+    'wss://picasso-rpc.composable.finance',
+  ],
 });
 
 export const picassoAlphanet = new Parachain({
@@ -1866,7 +1904,7 @@ export const picassoAlphanet = new Parachain({
   nativeAsset: pica,
   parachainId: 2019,
   ss58Format: 49,
-  ws: 'wss://boot-01.picasso2270.composablenodes.tech/',
+  ws: ['wss://boot-01.picasso2270.composablenodes.tech/'],
 });
 
 export const polkadot = new Parachain({
@@ -1879,7 +1917,11 @@ export const polkadot = new Parachain({
   nativeAsset: dot,
   parachainId: 0,
   ss58Format: 0,
-  ws: 'wss://polkadot-rpc.dwellir.com',
+  ws: [
+    'wss://polkadot-rpc.dwellir.com',
+    'wss://polkadot.api.onfinality.io/public-ws',
+    'wss://rpc.polkadot.io/',
+  ],
 });
 
 export const polkadotAssetHub = new Parachain({
@@ -1938,7 +1980,11 @@ export const polkadotAssetHub = new Parachain({
   nativeAsset: dot,
   parachainId: 1000,
   ss58Format: 42,
-  ws: 'wss://polkadot-asset-hub-rpc.polkadot.io',
+  ws: [
+    'wss://asset-hub-polkadot-rpc.dwellir.com',
+    'wss://polkadot-asset-hub-rpc.polkadot.io',
+    'wss://statemint.api.onfinality.io/public-ws',
+  ],
 });
 
 export const robonomics = new Parachain({
@@ -1955,7 +2001,10 @@ export const robonomics = new Parachain({
   nativeAsset: xrt,
   parachainId: 2048,
   ss58Format: 32,
-  ws: 'wss://kusama.rpc.robonomics.network/',
+  ws: [
+    'wss://kusama.rpc.robonomics.network/',
+    'wss://robonomics-rpc.dwellir.com',
+  ],
 });
 
 export const shiden = new Parachain({
@@ -1978,7 +2027,11 @@ export const shiden = new Parachain({
   nativeAsset: sdn,
   parachainId: 2007,
   ss58Format: 5,
-  ws: 'wss://shiden-rpc.dwellir.com',
+  ws: [
+    'wss://shiden-rpc.dwellir.com',
+    'wss://rpc.shiden.astar.network',
+    'wss://shiden.api.onfinality.io/public-ws',
+  ],
 });
 
 export const subsocial = new Parachain({
@@ -1991,7 +2044,7 @@ export const subsocial = new Parachain({
   nativeAsset: sub,
   parachainId: 2101,
   ss58Format: 28,
-  ws: 'wss://para.subsocial.network',
+  ws: ['wss://para.subsocial.network', 'wss://subsocial-rpc.dwellir.com'],
 });
 
 export const tinkernet = new Parachain({
@@ -2011,7 +2064,7 @@ export const tinkernet = new Parachain({
   nativeAsset: tnkr,
   parachainId: 2125,
   ss58Format: 117,
-  ws: 'wss://tinkernet-rpc.dwellir.com',
+  ws: ['wss://tinkernet-rpc.dwellir.com'],
 });
 
 export const turing = new Parachain({
@@ -2034,7 +2087,7 @@ export const turing = new Parachain({
   nativeAsset: tur,
   parachainId: 2114,
   ss58Format: 51,
-  ws: 'wss://rpc.turing.oak.tech',
+  ws: ['wss://rpc.turing.oak.tech', 'wss://turing-rpc.dwellir.com'],
 });
 
 export const turingAlphanet = new Parachain({
@@ -2058,7 +2111,7 @@ export const turingAlphanet = new Parachain({
   nativeAsset: tur,
   parachainId: 2114,
   ss58Format: 51,
-  ws: 'wss://turing-moonbase.rpc.oak.tech/',
+  ws: ['wss://turing-moonbase.rpc.oak.tech/'],
 });
 
 export const uniqueAlpha = new Parachain({
@@ -2079,7 +2132,7 @@ export const uniqueAlpha = new Parachain({
   nativeAsset: auq,
   parachainId: 2095,
   ss58Format: 255,
-  ws: 'wss://unique-alpha.unique.network',
+  ws: ['wss://unique-alpha.unique.network'],
 });
 
 export const zeitgeist = new Parachain({
@@ -2127,7 +2180,6 @@ export const chainsList: AnyChain[] = [
   crustShadow,
   darwinia,
   darwiniaCrab,
-  equilibrium,
   hydration,
   hydrationAlphanet,
   integritee,